+++ services/prisma.ts
@@ -7,25 +7,6 @@ try {
     log: ["info", "warn", "error"],
   });
   macros.log("** Created new Prisma client");
-  // TEMP / TODO - REMOVE / DO NOT LEAVE HERE PLEASE
-  // Temp fix to address Prisma connection pool issues
-  // https://github.com/prisma/prisma/issues/7249#issuecomment-1059719644
-  const intervalTime = 6 * 60 * 60_000; // Every 6 hours
-
-  if (!macros.TEST) {
-    setInterval(async () => {
-      const startTime = Date.now();
-      await prisma.$disconnect();
-      macros.log("Disconnected Prisma");
-      await prisma.$connect();
-      const totalTime = Date.now() - startTime;
-      macros.log(
-        `Reconnected Prisma - downtime of ${totalTime} ms (${
-          totalTime / 60_000
-        } mins)`
-      );
-    }, intervalTime);
-  }
 } catch (e) {
   macros.error(e);
 }
